test: исправлены тесты и устранены баги

+++ ff-tvm/internal/models/models.go
@@ -6,7 +6,7 @@ import (
 
 // Service представляет информацию о сервисе в БД
 type Service struct {
-	ID        int64  `gorm:"primaryKey"`
+	ID        int    `gorm:"primaryKey"`
 	Name      string `gorm:"not null"`
 	PublicKey string `gorm:"not null"`
 	CreatedAt time.Time
@@ -15,17 +15,17 @@ type Service struct {
 
 // ServiceAccess представляет разрешения доступа между сервисами
 type ServiceAccess struct {
-	ID        int64 `gorm:"primaryKey"`
-	FromID    int64 `gorm:"not null"`
-	ToID      int64 `gorm:"not null"`
+	ID        int `gorm:"primaryKey"`
+	FromID    int `gorm:"not null"`
+	ToID      int `gorm:"not null"`
 	CreatedAt time.Time
 	UpdatedAt time.Time
 }
 
 // KeyPair представляет пару ключей ED25519
 type KeyPair struct {
-	ID         int64  `gorm:"primaryKey"`
-	ServiceID  int64  `gorm:"not null"`
+	ID         int    `gorm:"primaryKey"`
+	ServiceID  int    `gorm:"not null"`
 	PublicKey  string `gorm:"not null"`
 	PrivateKey string `gorm:"not null"`
 	CreatedAt  time.Time
